+++ src/irina/config.rs
@@ -10,6 +10,8 @@ pub struct Config {
     pub email_password: String,
     pub imap_server: String,
     pub imap_port: u16,
+    pub redis_url: String,
+    pub rsshub_url: String,
     pub teloxide_token: String,
 }