+++ src/irina/youtube.rs
@@ -2,7 +2,8 @@
 //!
 //! This module exposes the function to fetch the youtube latest videos from spazio grigio
 
-use crate::youtube::{Feed, Video, YoutubeClient};
+use crate::feed::Entry;
+use crate::youtube::{Feed, YoutubeClient};
 
 // <https://www.youtube.com/feeds/videos.xml?channel_id=UCK3cMi97Kf_WENLvRFdztoQ>
 const CHANNEL_ID: &str = "UCK3cMi97Kf_WENLvRFdztoQ";
@@ -11,8 +12,8 @@ pub struct Youtube;
 
 impl Youtube {
     /// Get latest video from spazio grigio
-    pub async fn get_latest_video() -> anyhow::Result<Video> {
-        if let Some(video) = Self::get_latest_videos().await?.videos().next() {
+    pub async fn get_latest_video() -> anyhow::Result<Entry> {
+        if let Some(video) = Self::get_latest_videos().await?.entries().next() {
             Ok(video.clone())
         } else {
             anyhow::bail!("Ciao sono Irina. Non ho trovato nessun video per te")