+++ src/main.rs
@@ -4,15 +4,16 @@ const APP_VERSION: &str = env!("CARGO_PKG_VERSION");
 #[macro_use]
 extern crate lazy_regex;
 #[macro_use]
-extern crate lazy_static;
-#[macro_use]
 extern crate serde;
 #[macro_use]
 extern crate tracing;
 
+mod feed;
 mod irina;
 mod mail;
+mod redis;
 mod repository;
+mod rsshub;
 mod utils;
 mod youtube;