+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.2.0](#020)
   - [0.1.1](#011)
   - [0.1.0](#010)
 
 ---
+
+## 0.2.0
+
+Released on 25/08/2022
+
+- Added redis server to track fetch dates
+- Removed startup notify
+- Added rsshub to fetch instagram posts
 
 ## 0.1.1